fix: circular dependency
CreateTokenTransfer was importing types from index.tsx which was importing types from CreateTokenTransfer. This lead to a circular dependencies and values turned out to be undefined in the code
