+++ apps/web/src/components/common/TokenAmountInput/index.tsx
@@ -8,14 +8,13 @@ import { useCallback } from 'react'
 import { get, useFormContext } from 'react-hook-form'
 import type { FieldArrayPath, FieldValues } from 'react-hook-form'
 import css from './styles.module.css'
-import { MultiTokenTransferFields, type MultiTokenTransferParams } from '@/components/tx-flow/flows/TokenTransfer'
+import {
+  MultiTokenTransferFields,
+  type MultiTokenTransferParams,
+  TokenAmountFields,
+} from '@/components/tx-flow/flows/TokenTransfer/types'
 import { sameAddress } from '@safe-global/utils/utils/addresses'
 import { type Balances } from '@safe-global/store/gateway/AUTO_GENERATED/balances'
-
-export enum TokenAmountFields {
-  tokenAddress = 'tokenAddress',
-  amount = 'amount',
-}
 
 export const InsufficientFundsValidationError = 'Insufficient funds'
