+++ apps/web/src/components/tx-flow/flows/TokenTransfer/RecipientRow/index.tsx
@@ -4,8 +4,8 @@ import DeleteIcon from '@/public/images/common/delete.svg'
 import { Box, Button, FormControl, Stack, SvgIcon } from '@mui/material'
 import { get, useFormContext } from 'react-hook-form'
 import type { FieldArrayPath, FieldPath } from 'react-hook-form'
-import type { MultiTokenTransferParams, TokenTransferParams } from '..'
-import { MultiTokenTransferFields, TokenTransferFields, TokenTransferType } from '..'
+import type { MultiTokenTransferParams, TokenTransferParams } from '../types'
+import { MultiTokenTransferFields, TokenTransferFields, TokenTransferType } from '../types'
 import { useTokenAmount } from '../utils'
 import { useHasPermission } from '@/permissions/hooks/useHasPermission'
 import { Permission } from '@/permissions/config'
