+++ apps/web/src/components/tx-flow/flows/TokenTransfer/index.tsx
@@ -2,40 +2,20 @@ import CreateTokenTransfer from './CreateTokenTransfer'
 import ReviewTokenTx from '@/components/tx-flow/flows/TokenTransfer/ReviewTokenTx'
 import AssetsIcon from '@/public/images/sidebar/assets.svg'
 import { ZERO_ADDRESS } from '@safe-global/protocol-kit/dist/src/utils/constants'
-import { TokenAmountFields } from '@/components/common/TokenAmountInput'
 import { useMemo } from 'react'
 import { TxFlowType } from '@/services/analytics'
 import { TxFlow } from '../../TxFlow'
 import { TxFlowStep } from '../../TxFlowStep'
+import { TokenTransferType, type MultiTokenTransferParams, type TokenTransferParams } from './types'
 
-export enum TokenTransferType {
-  multiSig = 'multiSig',
-  spendingLimit = 'spendingLimit',
-}
-
-enum Fields {
-  recipient = 'recipient',
-}
-
-export const TokenTransferFields = { ...Fields, ...TokenAmountFields }
-
-export type TokenTransferParams = {
-  [TokenTransferFields.recipient]: string
-  [TokenTransferFields.tokenAddress]: string
-  [TokenTransferFields.amount]: string
-}
-
-export enum MultiTransfersFields {
-  recipients = 'recipients',
-  type = 'type',
-}
-
-export const MultiTokenTransferFields = { ...MultiTransfersFields }
-
-export type MultiTokenTransferParams = {
-  [MultiTransfersFields.recipients]: TokenTransferParams[]
-  [MultiTransfersFields.type]: TokenTransferType
-}
+export {
+  TokenTransferFields,
+  TokenTransferType,
+  MultiTransfersFields,
+  MultiTokenTransferFields,
+  type TokenTransferParams,
+  type MultiTokenTransferParams,
+} from './types'
 
 type MultiTokenTransferFlowProps = {
   recipients?: Partial<TokenTransferParams>[]
